use materialized view materialization instead of table (#9)
We are currently hacking the "table" materialization to create a materialized view
under the hood. While this is still useful for testing purposes, it turns out that we can
just let users indicate "materializedview" as the materialization strategy for their
models, instead. This is a UX change that maps dbt + Materialize usage to what users
would be executing in Materialize directly.